12.2.Calling the Controller & View Event{The kernel.view Event}
Go back and remove that.

At this point, we definitely have a Response object: either our controller returned a Response or a listener to kernel.view was able to create one. So finally, down here... we're done! We made it! We return $this->filterResponse($response). What does that method do? Do you remember what you're always supposed to answer when I ask that question? Yep, it dispatches another event! Let's look into that next and dig into a few really cool listeners on it, including one that takes us into the topic of the "request format".