15.2.Exception Handling{Setting Response, Stop Propagation}
Back in the controller, remove that code.

Setting Response, Stop Propagation
Head over to HttpKernel. Symfony ultimately wants a Response object: it wants somebody to set the response on this event, which it fetches with $response = $event->getResponse().
Hold Command or Ctrl and click to open the ExceptionEvent class. This class is similar to the RequestEvent that we saw earlier. If you find handleRaw()... here it is: RequestEvent is used for the earliest event in Symfony. Listeners to that event are able to set a Response if they want.
The same happens down in handleThrowable(): listeners can set a Response on the event. In fact, in ExceptionEvent, check out the base class! It's RequestEvent! It's the exact class the other event uses, and it holds the setResponse() method.
We already saw that method used in RouterListener: $event->setResponse() with $this->createWelcomeResponse().
But what I really want to show you, back in RequestEvent, is this: if something calls setResponse(), the event class calls a stopPropagation() method. If you hold Command or Ctrl to jump to that, it opens in another base class. This method sets a flag called propagationStopped to true.
This is important: if you have multiple listeners to kernel.exception and one of them sets the Response, the other listeners will not be called. Yep, the EventDispatcher looks for this flag and, if it's true, it immediately stops calling the other listeners. This means that the first listener to set a response wins. It's a good thing to keep in mind, and it explains some of the priorities that the listeners have.
Next: let's look at the critical listeners to the kernel.exception event.